requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.8
colorama==0.4.6
dnspython==1.16.0
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
motor==3.7.0
psycopg2==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==3.11.0
python-decouple==3.8
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
